metro-prediction-api/requirement.txt
# Web框架
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1

# 数据库
pymysql==1.1.0
SQLAlchemy==2.0.23

# 数据处理
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# 机器学习/AI
onnxruntime==1.16.3

# 工具库
python-dotenv==1.0.0
marshmallow==3.21.0
schedule==1.2.0
APScheduler==3.10.4

# 部署
gunicorn==21.2.0

# 开发工具
pytest==7.4.3
black==23.11.0
flake8==6.1.0